Add check for log file getting smaller
As it was, if the log file was truncated (inode, devices, etc unchanged)
logsufer didn't notice and kept reading 0 bytes at it's current file
position, now way past the end of the file (at least, until the file
becomes bigger than it was after the last read before the truncate).

Added a check for file size becoming smaller, so that if the log is
truncated (assuming it doesn't become larger than it was after the
last read before the truncate) then it will be reopened.