Single match_data block.
The match_data block contains details of the matching string and
captured substrings of a match. There is limited use of these and
due to constraints elsewhere in the code a maximum of RE_NREGS
sub matches are supported anyway, so there is no value in allocating
match_block according to the regular expression to accommodate more.